Finished Data Adding Buttons
Added 3 Buttons for adding items to JSON file
Adding a Key, Value
Adding a JSON Array
Adding a JSON Object

Items will be added to selected object if its an Array or JSON object, or to the Parent Object if the selected item is just a JSON Value.
Need to explain how item selection and placement works in the README most likely or in some sort of tutorial or help menu.